requirements.txt
amqp==1.4.6
anyjson==0.3.3
Babel==1.3
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.0
billiard==3.3.0.20
boto==2.38.0
celery==3.1.18
certifi==2015.4.28
click==4.0
cssmin==0.2.0
DateTime==4.0.1
dj-database-url==0.3.0
dj-redis-url==0.1.4
dj-static==0.0.6
django-allauth==0.21.0
django-angular==0.7.15
django-appconf==1.0.1
django-compressor==1.5
django-countries==3.3
django-debug-toolbar==1.3.2
django-filter==0.10.0
django-hashers-passlib==0.2
django-modeltranslation==0.10
django-mptt==0.7.4
django-nested-inline==0.3.4
django-nodebow==0.0.1
django-redis-cache==1.1.0
django-redis-sessions-fork==0.6.2
django-rest-auth==0.4.0
django-rosetta==0.7.6
django-storages==1.1.8
django-timedeltafield==0.7.10
django-timezone-field==1.2
django-toolbelt==0.0.1
django-watson==1.1.8
django-websocket-redis==0.4.4
django-wysiwyg-redactor==0.4.8.1
Django==1.8.2
djangorestframework==3.1.3
ecdsa==0.13
Fabric==1.10.2
flower==0.8.2
futures==3.0.3
gevent==1.1a1
greenlet==0.4.7
gspread==0.2.5
gunicorn==19.3.0
hiredis==0.2.0
html5lib==1.0b7
httplib2==0.9.1
jsmin==2.1.1
jsonfield==1.0.3
kombu==3.0.26
microsofttranslator==0.7
newrelic==2.52.0.40
nodeenv==0.13.2
oauth2client==1.4.11
oauthlib==0.7.2
paramiko==1.15.2
passlib==1.6.2
Pillow==2.9.0
pip-tools==0.3.6
polib==1.0.6
psycopg2==2.6.1
pyasn1-modules==0.0.6
pyasn1==0.1.8
pycrypto==2.6.1
PyExecJS==1.1.0
python-dateutil==2.4.2
python-dotenv==0.1.2
python-openid==2.2.5
python-postmark==0.4.7
pytz==2015.4
redis==2.10.3
requests-oauthlib==0.5.0
requests==2.7.0
rsa==3.1.4
six==1.9.0
socketIO-client==0.6.5
sorl-thumbnail==12.2
sqlparse==0.1.15
static3==0.6.1
suds==0.4
threadpool==1.2.7
tornado==4.2
uWSGI==2.0.11
websocket-client==0.32.0
zope.interface==4.1.2
